website — копия/core/requirements.txt
alembic==1.3.2
aniso8601==9.0.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.1.7
bleach==3.3.0
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
click==7.1.2
cycler==0.11.0
decorator==5.1.1
distlib==0.3.5
Django==4.0.6
django-js-asset==2.0.0
django-mptt==0.13.4
dnspython==2.2.1
email-validator==1.1.3
executing==0.9.1
filelock==3.7.1
Flask==1.1.4
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
flask-restx==0.4.0
Flask-Script==2.0.6
Flask-Security-Too==4.0.1
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.34.4
gunicorn==20.1.0
idna==2.8
ipython==8.4.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
joblib==1.1.0
jsonschema==4.7.2
kiwisolver==1.4.4
Mako==1.2.1
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nltk==3.7
numpy==1.23.1
packaging==21.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyMySQL==0.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2022.1
PyYAML==5.3.1
regex==2022.7.25
requests==2.22.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.12
sqlparse==0.4.2
stack-data==0.3.0
task-manager==6.4.0
tqdm==4.64.0
traitlets==5.3.0
urllib3==1.25.11
virtualenv==20.15.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==6.2.0
WTForms==3.0.1
wtforms-html5==0.6.1
wxPython==4.1.1
